#3 wip layout cutting
