+++ src/App.js
@@ -1,11 +1,26 @@
-import React from 'react';
-import Menu from './components/Menu';
-import './Styles/App.css';
+import React from "react";
+import { Layout } from "antd";
+import SideBar from "./components/SideBar";
+import TopBar from "./components/TopBar";
+import "./Styles/App.css";
+const { Content } = Layout;
 
-function App () {
+function App() {
   return (
     <div className='App'>
-      <Menu />
+      <Layout style={{ minHeight: "100vh" }}>
+        <SideBar />
+        <Layout className='site-layout'>
+          <TopBar />
+          <Content style={{ margin: "0 16px" }}>
+            <div
+              className='site-layout-background'
+              style={{ padding: 24, minHeight: 360 }}>
+              Bill is a cat.
+            </div>
+          </Content>
+        </Layout>
+      </Layout>
     </div>
   );
 }
