+++ src/components/TopBar.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { Layout } from "antd";
+
+const { Header, Breadcrumb } = Layout;
+
+const TopBar = () => {
+  return (
+    <>
+      <Header className='site-layout-background' style={{ padding: 0 }} />
+    </>
+  );
+};
+
+export default TopBar;
